Update manifest.json and add Twitter redirects
The Twitter function is written in a content_script which is
called as soon as possible and then rewrites window.location
with the proper Nitter instance. It ignores
webRequest.onBeforeRequest, the backbone of the Chrome and
FF extension, which is not available in Safari.

Support for other sites will be built in this same,
compartmental fashion – adding the proper domains to the
manifest.json file and then creating a self-contained JS
file to do the redirect.