Replace popup with button to open the native app
The popup was previously a list of quick toggles, which have already
been moved over the native application. This is the next step in the
migration of the extension's settings away from the browser's storage
and into the native application.

Opening the app from the popup is incredibly complicated. I spent my
entire day scouring the internet for documentation, and rewrote the
implementation several times until I finally cracked it. I very well
could be one of the only people who doesn't work at Apple who has got
this working (it helps that there are few developers working on
extensions for Safari at the moment).

The process takes a couple steps:
1. Whenever the native app opens, it checks the "location" key in a
   shared UserDefaults suite (defined in the App Groups capability of
   both the extension and the native app). If the key is undefined, it
   sets it to "Bundle.main.bundleURL" – the location of the native app
   on disk.
2. From Safari, the user clicks on an "Open App" button in the popup,
   which fires off an "Open Sesame" message to the background script via
   the sendMessage API.
3. The background script, listening for messages, sends another "Open
   Sesame" message, this time from the "web extension" to the "app
   extension" using the sendNativeMessage API.
4. The "app extension" (an intermediary between the native app and web
   extension) then loads the "location" key from the shared UserDefaults
   suite, and checks the message it's received to see if it's an "Open
   Sesame" message. If it is, it opens the native app with
   NSWorkspace.shared.openApplication, using the value of the "location"
   key to set the app to open.

Note that "Open Sesame" can be any arbitrary string.

Next Steps:
1. Make the popup look decent (now it's just a button)
2. Adapt the web extension to native app communication to modify
   settings