Update app extension to send settings on request
This update changes the behavior of the Swift "app extension," which now
returns the dictionary representation of UserDefaults on a settings
request. This approach may change in the future to limit the amount of
data sent.

The second half of this process will be making the JavaScript web
extension side of Privacy Redirect for Safari receive the settings. This
is quite complicated, as it involves a number of Promises and message
between processes. Hopefully it will be done soon, because it is
starting to get on my nerves.