Update version strings
I've developed a new understanding of how these work, though the
documentation is pretty unclear. CFBundleShortVersionString is the
public release version (1.0.0), and increments linearly. CFBundleVersion
on the other hand is a (arbitrary?) string that can change without
incrementing the release version.

What this means in practice is that the version-incrementing problem
I've experienced in the past with App Review is not an issue. If an app
is rejected, I don't need to increase the release version, just the
build version. Yay.