Add fallbacks for undefined UserDefaults keys
I'm still not certain with the defaults in our @AppStorage property
wrappers are not properly set, but I'm leaning towards this being a bug
in Swift(UI).

This workaround is transparent on the JavaScript side (i.e. requires no
changes), and simply adds defaults as fallbacks on the Swift side of
things if the settings don't exist.

Closes #2.