Replace launch storyboard with blank launch screen
The iOS app previously had a "UILaunchStoryboardName" key with an empty
string where a file was supposed to be. Storyboards are stupid, however,
because they are graphical and annoying and not declarative like
SwiftUI. I have thus deleted this and replaced it with an empty
UILaunchScreen in the Info.plist file.

This also makes it so that the iOS app passes all validation checks for
the App Store.