Redesign app icons with Sketch
Due to repeated rejections on the iOS App Store (though not macOS)
caused by the use of "Apple trademark imagery - or likeness - in the the
app icon," we have had to tweak the logo slightly.

The new design replaces the Safari logo at the center with a more
general globe with meridians, symbolizing the internet.

We've also taken the time to, with this change, move our icons out of
SVG format and into Sketch. As much as I would like to keep them in an
open format, the editing experience provided by Sketch is leagues ahead
of that of Inkscape.